+++ src/main/java/dev/j3fftw/litexpansion/Events.java
@@ -1,6 +1,10 @@
 package dev.j3fftw.litexpansion;
 
+import org.bukkit.Bukkit;
+import org.bukkit.Location;
+import org.bukkit.Material;
 import org.bukkit.Sound;
+import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.entity.Player;
@@ -10,7 +14,10 @@ import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
 import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.FoodLevelChangeEvent;
+import org.bukkit.event.player.PlayerInteractEvent;
+import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.ItemStack;
+
 
 import dev.j3fftw.litexpansion.armor.ElectricChestplate;
 import dev.j3fftw.litexpansion.items.FoodSynthesizer;
@@ -18,8 +25,11 @@ import dev.j3fftw.litexpansion.items.Wrench;
 import dev.j3fftw.litexpansion.utils.Constants;
 import dev.j3fftw.litexpansion.utils.Utils;
 import dev.j3fftw.litexpansion.weapons.NanoBlade;
+import io.github.thebusybiscuit.slimefun4.implementation.items.cargo.TrashCan;
 import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetComponent;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.interfaces.InventoryBlock;
+import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
 import me.mrCookieSlime.Slimefun.cscorelib2.protection.ProtectableAction;
@@ -95,10 +105,91 @@ public class Events implements Listener {
 
             if (slimefunBlock instanceof EnergyNetComponent) {
                 e.setCancelled(true);
-                Wrench.wrenchBlock(p, block, true, false);
+                wrenchBlock(p, block, true, false);
                 Utils.send(p, "&cYou need a Wrench to break Slimefun machines!");
                 Utils.send(p, "&c(Slimefun Guide > LiteXpansion > Wrench)");
             }
+        }
+    }
+
+    @EventHandler
+    @SuppressWarnings("ConstantConditions")
+    public void onWrenchUse(PlayerInteractEvent e) {
+
+        Player p = e.getPlayer();
+        Wrench wrench = (Wrench) Items.WRENCH.getItem();
+        ItemStack playerWrench = p.getInventory().getItemInMainHand();
+        final Block block = e.getClickedBlock();
+
+        if (block == null) return;
+
+        if (e.getHand() == EquipmentSlot.HAND && wrench.isItem(playerWrench)
+            && SlimefunPlugin.getProtectionManager().hasPermission(e.getPlayer(),
+            block.getLocation(), ProtectableAction.BREAK_BLOCK)) {
+
+            SlimefunItem sfBlock = BlockStorage.check(block);
+
+            if (sfBlock instanceof EnergyNetComponent) {
+
+                if (Constants.MACHINE_BREAK_REQUIRES_WRENCH) {
+
+                    double random = Math.random();
+                    wrenchBlock(p, block, random <= Constants.WRENCH_FAIL_CHANCE, true);
+
+                } else {
+                    wrenchBlock(p, block, false, true);
+                }
+
+                wrench.damageItem(p, playerWrench);
+
+            } else if (sfBlock != null && (sfBlock.getID().startsWith("CARGO_NODE")
+                || sfBlock instanceof TrashCan)) {
+
+                wrenchBlock(p, block, false, true);
+                wrench.damageItem(p, playerWrench);
+
+            } else {
+                Utils.send(p, "&cYou can not use the wrench on this block!");
+            }
+        }
+    }
+
+    public static void wrenchBlock(Player p, Block block, boolean fail, boolean byInteract) {
+
+        if (fail) {
+
+            World blockWorld = block.getWorld();
+            Location blockLocation = block.getLocation();
+            SlimefunItem slimefunBlock = BlockStorage.check(block);
+            BlockMenu blockInventory = BlockStorage.getInventory(block);
+
+            if (BlockStorage.hasInventory(block)) {
+                int[] inputSlots = ((InventoryBlock) slimefunBlock).getInputSlots();
+                for (int slot : inputSlots) {
+
+                    if (blockInventory.getItemInSlot(slot) != null) {
+                        blockWorld.dropItemNaturally(blockLocation, blockInventory.getItemInSlot(slot));
+                    }
+                }
+                int[] outputSlots = ((InventoryBlock) slimefunBlock).getOutputSlots();
+                for (int slot : outputSlots) {
+                    if (blockInventory.getItemInSlot(slot) != null) {
+                        blockWorld.dropItemNaturally(blockLocation, blockInventory.getItemInSlot(slot));
+                    }
+                }
+            }
+
+            BlockStorage.clearBlockInfo(block);
+            block.setType(Material.AIR);
+            block.getWorld().dropItemNaturally(block.getLocation(), Items.MACHINE_BLOCK.clone());
+
+            if (byInteract) {
+                Utils.send(p, "&cOh no! Your wrench failed!");
+            }
+
+        } else {
+            BlockBreakEvent breakEvent = new BlockBreakEvent(block, p);
+            Bukkit.getServer().getPluginManager().callEvent(breakEvent);
         }
     }
